set up useWordCheck hook
