+++ src/helpers/index.js
@@ -1,3 +1,4 @@
 import getKeyClassNames from "./getKeyClassNames";
+import checkLetter from "./checkLetter";
 
-export { getKeyClassNames };
+export { getKeyClassNames, checkLetter };
