+++ src/hooks/index.js
@@ -1,0 +1,3 @@
+import useWordCheck from "./useWordCheck";
+
+export { useWordCheck };
